requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
sounddevice>=0.4.6
pyaudio>=0.2.13

# Data Processing
scikit-learn>=1.3.0
pandas>=2.0.0

# Raspberry Pi GPIO
RPi.GPIO>=0.7.1

# Utilities
pynput>=1.7.6
tqdm>=4.65.0

# Optional: Model Visualization
# torchviz>=0.0.2
# torchsummary>=1.5.1

